+++ server/CalcServer/Services/ICalculator.cs
@@ -7,6 +7,6 @@ namespace CalcServer.Services
         /// <summary> Gets the solution for a given calculation command. </summary>
         /// <param name="command"> Command to execute. </param>
         /// <returns> The result of execution. </returns>
-        int Solve(CalculationCommand command);
+        float Solve(CalculationCommand command);
     }
 }
